Add ability to put the compiled fex file into the SD image
There is a new recipe called sunxi-board-fex which cares about the
compiling of the FEX file for each machine. If there is a fex.bin file
available in the deploy directory and machine depends on it (see SD
image class) it will be added.

Only OLinuXino A10S uses it for now.

Signed-off-by: Tomas Novotny <novotny@rehivetech.com>