README.md: improve instructions for sunxi versions
If the user wants to build the sunxi version of the kernel they can specify
the PREFERRED_PROVIDER block in their conf/local.conf (i.e. they don't need to
edit the layer explicitly).

Also, now that gcc-5.2 is the default on master, inform the user how they can
continue to build the sunxi version (3.4%) version of the kernel by forcing
the compiler to version 4.9.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>