machine: sunxi: remove UBOOT_LOADADDRESS
This has been assigned a value containing a typo since 2015, my
assumption is that this is not used and hence we can simply drop
it. The typo is that the address contains a 'O' character instead
of '0', the last one is a zero.

Either way the kernel.bbclass will set a default value, that is:

     UBOOT_LOADADDRESS ?= "${UBOOT_ENTRYPOINT}"

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>